refactor!: upgrade aiogram and pydantic

+++ yakusoku/modules/waifu/main.py
@@ -1,16 +1,19 @@
 import contextlib
 import html
+import time
 import traceback
 from dataclasses import dataclass
 from datetime import datetime, timedelta
 
-from aiogram.dispatcher.filters import ChatTypeFilter
+from aiogram import Bot
+from aiogram.dispatcher.event.bases import SkipHandler
+from aiogram.enums import ChatAction, ChatMemberStatus
+from aiogram.filters import Command, CommandObject
+from aiogram.filters.callback_data import CallbackData, CallbackQueryFilter
 from aiogram.types import (
+    BufferedInputFile,
     CallbackQuery,
-    ChatActions,
-    ChatMemberStatus,
     ChatMemberUpdated,
-    ChatType,
     InlineKeyboardButton,
     InlineKeyboardMarkup,
     Message,
@@ -21,7 +24,7 @@ from yakusoku.archive import utils as archive_utils
 from yakusoku.archive.exceptions import ChatDeleted, ChatNotFound
 from yakusoku.archive.models import UserData
 from yakusoku.context import common_config, module_manager, sql
-from yakusoku.filters import CallbackQueryFilter, ManagerFilter, NonAnonymousFilter
+from yakusoku.filters import GroupFilter, ManagerFilter, NonAnonymousFilter
 from yakusoku.shared.callback import CallbackQueryTaskManager
 from yakusoku.shared.lock import SimpleLockManager
 from yakusoku.utils import chat, exception
@@ -37,11 +40,11 @@ from .manager import (
 from .models import WAIFU_DEFAULT_RARITY, WAIFU_MAX_RARITY, WAIFU_MIN_RARITY
 from .registry import Registry
 
-dp = module_manager.dispatcher()
+router = module_manager.create_router()
 
 _manager = WaifuManager(sql)
 _registry = Registry(_manager)
-_tasks = CallbackQueryTaskManager(dp, "waifu_task/", "任务不见力 QwQ")
+_tasks = CallbackQueryTaskManager(router, "waifu_task/", "任务不见力 QwQ")
 _registry_lock = SimpleLockManager()
 _graph_lock = SimpleLockManager()
 
@@ -52,27 +55,29 @@ class MemberWaifuInfo:
     waifu: int
 
 
-@dp.message_handler(
-    ChatTypeFilter([ChatType.GROUP, ChatType.SUPERGROUP]),  # type: ignore
-    NonAnonymousFilter(),
-    commands=["waifu"],
-)
-async def waifu(message: Message):
+class Propose(CallbackData, prefix="waifu_propose"):
+    first: int
+    second: int
+
+
+@router.message(Command("waifu"), GroupFilter, NonAnonymousFilter)
+async def waifu(message: Message, bot: Bot):
+    assert message.from_user and message.chat
+    user_id = message.from_user.id
+
     async def _get_waifu(
         message: Message, force: bool = False
     ) -> tuple[WaifuFetchResult, UserData]:
-        result = await _manager.fetch_waifu(message.chat.id, message.from_id, force)
+        result = await _manager.fetch_waifu(message.chat.id, user_id, force)
         try:
-            waifu = await archive_utils.fetch_member(
-                message.bot, message.chat.id, result.waifu, True
-            )
+            waifu = await archive_utils.fetch_member(bot, message.chat.id, result.waifu, True)
             return (result, waifu)
         except ChatDeleted:
             if result.state == WaifuFetchState.RESTRICTED:
-                await _registry.divorce(message.chat.id, message.from_id)
+                await _registry.divorce(message.chat.id, user_id)
             return await _get_waifu(message, True)
 
-    await message.answer_chat_action(ChatActions.TYPING)
+    await bot.send_chat_action(message.chat.id, ChatAction.TYPING)
 
     try:
         result, waifu = await _get_waifu(message)
@@ -99,35 +104,31 @@ async def waifu(message: Message):
         InlineKeyboardMarkup(
             inline_keyboard=[
                 [
-                    InlineKeyboardButton(  # type: ignore
+                    InlineKeyboardButton(
                         text="成立婚事! (仅双方)",
-                        callback_data=f"waifu_propose_callback {message.from_id} {waifu.id}",
-                    ),
+                        callback_data=Propose(first=user_id, second=waifu.id).pack(),
+                    )
                 ]
             ]
         )
         if result.state != WaifuFetchState.RESTRICTED
-        else InlineKeyboardMarkup()
+        else None
     )
 
     with contextlib.suppress(Exception):
-        avatar = await avatar_manager.get_avatar_file(message.bot, waifu.id)
+        avatar = await avatar_manager.get_avatar_file(bot, waifu.id)
         if avatar:
             return await message.reply_photo(avatar.file_id, comment, reply_markup=buttons)
 
     await message.reply(comment, reply_markup=buttons)
 
 
-@dp.message_handler(
-    ChatTypeFilter([ChatType.GROUP, ChatType.SUPERGROUP]),  # type: ignore
-    ManagerFilter(),
-    commands=["waifurs"],
-)
-async def waifu_rarity_set(message: Message):
+@router.message(Command("waifurs"), GroupFilter, ManagerFilter)
+async def waifu_rarity_set(message: Message, command: CommandObject):
     if (
-        not (args := message.get_args())
+        not (args := command.args)
         or len(args := args.split()) != 2
-        or (rarity := exception.try_or_default(lambda: int(args[1]))) is None
+        or (rarity := exception.try_or_default(lambda: int(args[1]))) is None  # type: ignore
         or not (WAIFU_MIN_RARITY <= rarity <= WAIFU_MAX_RARITY)
     ):
         return await message.reply(
@@ -147,12 +148,9 @@ async def waifu_rarity_set(message: Message):
     await message.reply(f"成功将 {archive_utils.user_mention(waifu)} 的老婆稀有度修改为 {rarity}!")
 
 
-@dp.message_handler(
-    ChatTypeFilter([ChatType.GROUP, ChatType.SUPERGROUP]),  # type: ignore
-    commands=["waifurg"],
-)
+@router.message(Command("waifurg"), GroupFilter)
 async def waifu_rarity_get(message: Message):
-    if len(args := message.text.split()) != 2:
+    if not message.text or len(args := message.text.split()) != 2:
         return await message.reply("戳啦, 正确用法为 `/waifurg <@用户或ID>`", parse_mode="Markdown")
     try:
         waifu = await archive_utils.parse_member(
@@ -168,6 +166,14 @@ def create_divorce_task_unchecked(
     chat: int, originator: UserData, target: UserData
 ) -> InlineKeyboardMarkup:
     async def divorce(query: CallbackQuery):
+        if not isinstance(query.message, Message):
+            return await query.answer("消息太远古了, 我不是考古学家w")
+        if query.from_user.id == originator.id:
+            await query.answer("耐心等待对方接受哦w")
+            return False
+        if query.from_user.id != target.id:
+            await query.answer("别人的事情不要随便介入哦w")
+            return False
         await _registry.divorce(chat, originator.id)
         await query.message.reply(
             f"呜呜呜, {archive_utils.user_mention(originator)} "
@@ -178,38 +184,26 @@ def create_divorce_task_unchecked(
             await query.message.delete()
 
     async def cancelled(query: CallbackQuery):
+        if not isinstance(query.message, Message):
+            return await query.answer("消息太远古了, 我不是考古学家w")
         if query.from_user.id == originator.id:
             await query.answer("取消离婚申请成功捏, 以后要和谐相处哦~")
-        else:
+        elif query.from_user.id == target.id:
             await query.message.reply(
                 f"{archive_utils.user_mention(originator)} 离婚申请被取消捏, 以后要和谐相处哦~",
                 reply=False,
             )
+        else:
+            await query.answer("别人的事情不要随便介入哦w")
+            return False
         with contextlib.suppress(Exception):
             await query.message.delete()
 
     async def disposed():
         _registry_lock.unlock_all_unchecked((chat, originator.id), (chat, target.id))
 
-    task = _tasks.create_task(
-        divorce,
-        [
-            (lambda query: query.from_user.id != originator.id, "耐心等待对方接受哦w"),  # type: ignore
-            (lambda query: query.from_user.id == target.id, "别人的事情不要随便介入哦w"),  # type: ignore
-        ],
-        expired_after=timedelta(days=1),
-        disposed=disposed,
-    )
-    cancellation_task = _tasks.create_cancellation_task(
-        task,
-        [
-            (
-                lambda query: query.from_user.id in (originator.id, target.id),  # type: ignore
-                "别人的事情不要随便介入哦w",
-            )
-        ],
-        cancelled,
-    )
+    task = _tasks.create_task(divorce, expired_after=timedelta(days=1), disposed=disposed)
+    cancellation_task = _tasks.create_cancellation_task(task, cancelled)
     return InlineKeyboardMarkup(
         inline_keyboard=[
             [
@@ -226,16 +220,14 @@ def create_divorce_task_unchecked(
     )
 
 
-@dp.message_handler(
-    ChatTypeFilter([ChatType.GROUP, ChatType.SUPERGROUP]),  # type: ignore
-    NonAnonymousFilter(),
-    commands=["divorce"],
-)
+@router.message(Command("divorce"), GroupFilter, NonAnonymousFilter)
 async def divorce(message: Message):
-    data = await _manager.get_waifu_data(message.chat.id, message.from_id)
+    assert message.from_user
+    user_id = message.from_user.id
+    data = await _manager.get_waifu_data(message.chat.id, user_id)
     if not (partner := data.get_partner()):
         return await message.reply("啊? 身为单身狗, 离婚什么???")
-    originator = await user_manager.get_user(message.from_user.id)
+    originator = await user_manager.get_user(user_id)
     target = await user_manager.get_user(partner)
     if not _registry_lock.lock_all((message.chat.id, originator.id), (message.chat.id, target.id)):
         return await message.reply("你或者对方正在处理某些事项哦~")
@@ -250,6 +242,14 @@ def create_proposal_task_unchecked(
     chat: int, originator: UserData, target: UserData
 ) -> InlineKeyboardMarkup:
     async def marry(query: CallbackQuery):
+        if not isinstance(query.message, Message):
+            return await query.answer("消息太远古了, 我不是考古学家w")
+        if query.from_user.id == originator.id:
+            await query.answer("耐心等待对方接受哦w")
+            return False
+        if query.from_user.id != target.id:
+            await query.answer("别人的事情不要随便介入哦w")
+            return False
         await _registry.marry(chat, originator.id, target.id)
         await query.message.reply(
             f"恭喜 {archive_utils.user_mention(originator)} "
@@ -261,39 +261,27 @@ def create_proposal_task_unchecked(
             await query.message.delete()
 
     async def cancelled(query: CallbackQuery):
+        if not isinstance(query.message, Message):
+            return await query.answer("消息太远古了, 我不是考古学家w")
         if query.from_user.id == target.id:
             await query.message.reply(
                 f"呜呜呜, {archive_utils.user_mention(originator)} "
                 f"被 {archive_utils.user_mention(target)} 拒绝了w",
                 reply=False,
             )
-        else:
+        elif query.from_user.id == originator.id:
             await query.answer("已经帮你取消力~")
+        else:
+            await query.answer("别人的事情不要随便介入哦w")
+            return False
         with contextlib.suppress(Exception):
             await query.message.delete()
 
     async def disposed():
         _registry_lock.unlock_all_unchecked((chat, originator.id), (chat, target.id))
 
-    task = _tasks.create_task(
-        marry,
-        [
-            (lambda query: query.from_user.id != originator.id, "耐心等待对方接受哦w"),  # type: ignore
-            (lambda query: query.from_user.id == target.id, "别人的事情不要随便介入哦w"),  # type: ignore
-        ],
-        expired_after=timedelta(days=1),
-        disposed=disposed,
-    )
-    cancellation_task = _tasks.create_cancellation_task(
-        task,
-        [
-            (
-                lambda query: query.from_user.id in (originator.id, target.id),  # type: ignore
-                "别人的事情不要随便介入哦w",
-            )
-        ],
-        cancelled,
-    )
+    task = _tasks.create_task(marry, expired_after=timedelta(days=1), disposed=disposed)
+    cancellation_task = _tasks.create_cancellation_task(task, cancelled)
     return InlineKeyboardMarkup(
         inline_keyboard=[
             [
@@ -310,19 +298,19 @@ def create_proposal_task_unchecked(
     )
 
 
-@dp.message_handler(
-    ChatTypeFilter([ChatType.GROUP, ChatType.SUPERGROUP]),  # type: ignore
-    NonAnonymousFilter(),
-    commands=["propose"],
-)
+@router.message(Command("propose"), GroupFilter, NonAnonymousFilter)
 async def propose(message: Message):
-    if (length := len(args := message.text.split())) == 2:
+    assert message.from_user
+    user_id = message.from_user.id
+
+    args = message.text.split() if message.text else []
+    if (length := len(args)) == 2:
         try:
             target = await archive_utils.parse_member(args[1], message.chat.id, message.bot)
         except (ChatDeleted, ChatNotFound):
             return await message.reply("呜, 找不到你所提及的用户w")
     elif length == 1 and message.reply_to_message:
-        if message.reply_to_message.sender_chat:
+        if message.reply_to_message.sender_chat or not message.reply_to_message.from_user:
             return await message.reply("呜, 不能跟匿名用户结婚捏w")
         if message.reply_to_message.from_user.is_bot:
             return await message.reply("呜, 不能跟机器人结婚捏w")
@@ -332,19 +320,17 @@ async def propose(message: Message):
             "戳啦, 正确用法为 `/propose <@用户或ID (或回复某个用户的消息)>`", parse_mode="Markdown"
         )
 
-    if message.from_id == target.id:
+    if user_id == target.id:
         return await message.reply("啊? 这是可以选的吗? w")
     if (await _manager.get_waifu_data(message.chat.id, target.id)).get_partner() or (
-        await _manager.get_waifu_data(message.chat.id, message.from_id)
+        await _manager.get_waifu_data(message.chat.id, user_id)
     ).get_partner():
         return await message.reply("你或者对方已经结过婚捏, 不能向对方求婚诺w")
 
-    if not _registry_lock.lock_all(
-        (message.chat.id, message.from_id), (message.chat.id, target.id)
-    ):
+    if not _registry_lock.lock_all((message.chat.id, user_id), (message.chat.id, target.id)):
         return await message.reply("你或者对方正在处理某些事项哦~")
 
-    originator = await user_manager.get_user(message.from_user.id)
+    originator = await user_manager.get_user(user_id)
     buttons = create_proposal_task_unchecked(message.chat.id, originator, target)
     await message.reply(
         f"你向 {archive_utils.user_mention(target)} 发起了求婚邀请",
@@ -352,10 +338,12 @@ async def propose(message: Message):
     )
 
 
-@dp.callback_query_handler(CallbackQueryFilter("waifu_propose_callback"))
-async def propose_callback(query: CallbackQuery):  # type: ignore
-    first_id, second_id = map(int, query.data.split()[1:])
-    if query.from_user.id not in (first_id, second_id):
+@router.callback_query(CallbackQueryFilter(callback_data=Propose))
+async def propose_callback(query: CallbackQuery, callback_data: Propose):  # type: ignore
+    if not isinstance(query.message, Message):
+        return await query.answer("消息太远古了, 我不是考古学家w")
+    first_id, second_id = callback_data.first, callback_data.second
+    if query.from_user.id not in [first_id, second_id]:
         return await query.answer("别人的事情不要随便介入哦w")
     originator = await user_manager.get_user(query.from_user.id)
     target = await user_manager.get_user(second_id if originator.id == first_id else first_id)
@@ -369,7 +357,7 @@ async def propose_callback(query: CallbackQuery):  # type: ignore
         return await query.answer("你或者对方正在处理某些事项哦~")
     buttons = create_proposal_task_unchecked(query.message.chat.id, originator, target)
     await query.message.reply(
-        f"{chat.get_mention(query.from_user)} "
+        f"{chat.mention_html(query.from_user)} "
         f"向 {archive_utils.user_mention(target)} 发起了求婚邀请",
         reply_markup=buttons,
         reply=False,
@@ -377,9 +365,10 @@ async def propose_callback(query: CallbackQuery):  # type: ignore
     await query.answer()
 
 
-@dp.message_handler(NonAnonymousFilter(), commands=["waifum"])
+@router.message(Command("waifum"), NonAnonymousFilter)
 async def mention_global(message: Message):
-    config = await _manager.get_waifu_config(message.from_id)
+    assert message.from_user
+    config = await _manager.get_waifu_config(message.from_user.id)
     config.mentionable = not config.mentionable
     await _manager.update_waifu_config(config)
     await message.reply(
@@ -389,20 +378,19 @@ async def mention_global(message: Message):
     )
 
 
-@dp.message_handler(
-    ChatTypeFilter([ChatType.GROUP, ChatType.SUPERGROUP]),  # type: ignore
-    commands=["waifug", "waifu_graph"],
-)
-async def waifu_graph(message: Message):
+@router.message(Command("waifug", "waifu_graph"), GroupFilter)
+async def waifu_graph(message: Message, bot: Bot):
+    assert message.chat
     if not _graph_lock.lock(message.chat.id):
         return await message.reply("呜呜呜, 别骂了, 别骂了, 在画了www")
     reply = await message.reply_sticker(common_config.writing_sticker)
-    await message.answer_chat_action(ChatActions.TYPING)
+    await bot.send_chat_action(message.chat.id, ChatAction.TYPING)
     try:
         datas = await _manager.get_active_waifu_datas(message.chat.id)
         waifu_dict = {data.member: data.waifu for data in datas if data.waifu}
-        image = await graph.render(message.bot, waifu_dict, "png")
-        await message.reply_photo(image)
+        image = await graph.render(bot, waifu_dict, "png")
+        file = BufferedInputFile(image, f"waifug-{message.chat.id}-{time.time()}.png")
+        await message.reply_photo(file)
     except Exception as ex:
         await message.reply(f"喵呜……渲染失败捏. {html.escape(str(ex))}")
         traceback.print_exc()
@@ -410,17 +398,17 @@ async def waifu_graph(message: Message):
     await reply.delete()
 
 
-@dp.my_chat_member_handler()
-@dp.chat_member_handler()
-async def member_update(update: ChatMemberUpdated):
+@router.my_chat_member
+@router.chat_member
+async def member_update(update: ChatMemberUpdated, bot: Bot):
     if (member := update.new_chat_member).status not in (
         ChatMemberStatus.LEFT,
-        ChatMemberStatus.BANNED,
         ChatMemberStatus.KICKED,
     ):
         return
-    if member.user.id == member.bot.id:
+    if member.user.id == bot.id:
         with contextlib.suppress(Exception):
             return await _manager.remove_group(update.chat.id)
     with contextlib.suppress(KeyError):
         await _manager.remove_waifu(update.chat.id, member.user.id)
+    raise SkipHandler
